pyproject.toml
[project]
name = "ns-hotopic"
version = "0.1.0"
description = "Track hot topics from the NodeSeek homepage"
readme = "README.md"
authors = [
    { name = "Impersonality", email = "864921462@qq.com" }
]
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
]

[project.scripts]
ns-hotopic = "ns_hotopic:main"

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
